backend/requirements.txt
fastapi>=0.112.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
aiogram==3.4.1
python-dotenv>=1.0.1,<2.0.0
httpx>=0.27.0,<1.0.0
supabase>=2.5.0,<3.0.0
sqlmodel>=0.0.14,<0.1.0
pydantic>=2.4.1,<2.6.0
pydantic-settings>=2.2.1
PyJWT>=2.8.0,<3.0.0

